pyproject.toml
[tool.poetry]
name = "template_script"
version = "0.1.0"
description = "Template to execute scripts"
authors = ["Jônatas Luiz de Oliveira <contact@jonatasoliveira.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
SQLAlchemy = "^1.4.31"
mysql-connector-python = "^8.0.28"
httpx = "^0.22.0"
python-decouple = "^3.6"
pytest-cov = "^3.0.0"
pytest-watch = "^4.2.0"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.7.0"
pytest-asyncio = "^0.18.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
pytest = "^7.0.1"
isort = "^5.10.1"
blue = "^0.7.0"
darker = "^1.4.1"
mkdocs = "^1.2.3"
prospector = "^1.6.0"
aiosqlite = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
